Add files via upload
Added mobile controls, run and sprint toggles on and off without auto run, you activate the run or sprint and then when moving you will run or sprint but I have to repeat it, it has no autorun.  Also on PC you dodge by double pressing the movement keys very fast, on mobile you can either double press  the movement pads or slide your finger from left to right or right to left or up to down or down to up to dodge in the direction of the last pad.  so for example left to right means you dodge to the right, right to left means you dodge to the left.